Redundant codes

+++ Ultrapowa Clash Server/Helpers/GamePlayUtil.cs
@@ -1,15 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.IO;
-using System.Collections.Concurrent;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
-using UCS.GameFiles;
 using UCS.Core;
-using UCS.Logic;
+using UCS.GameFiles;
 
 namespace UCS.Helpers
 {
@@ -29,13 +20,12 @@ namespace UCS.Helpers
 
         public static int CalculateResourceCost(int sup, int inf, int supCost, int infCost, int amount)
         {
-            return (int)Math.Round((long)(supCost - infCost) * (long)(amount - inf) / (sup - inf * 1.0)) + infCost;
+            return (int) Math.Round((supCost - infCost)*(long) (amount - inf)/(sup - inf*1.0)) + infCost;
         }
 
         public static int CalculateSpeedUpCost(int sup, int inf, int supCost, int infCost, int amount)
         {
-            return (int)Math.Round((long)(supCost - infCost) * (long)(amount - inf) / (sup - inf * 1.0)) + infCost;
+            return (int) Math.Round((supCost - infCost)*(long) (amount - inf)/(sup - inf*1.0)) + infCost;
         }
     }
-
 }
